+++ ThirdPartyAdapters/inmobi/inmobi/src/test/kotlin/com/google/ads/mediation/inmobi/InMobiAdapterUtilsTest.kt
@@ -9,8 +9,10 @@ import com.inmobi.sdk.InMobiSdk
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 
@@ -78,7 +80,7 @@ class InMobiAdapterUtilsTest {
   }
 
   @Test
-  fun setIsAgeRestricted_whenCOPPANotSpecified_setsAgeRestrictedFalseOnInMobiSDK() {
+  fun setIsAgeRestricted_whenCOPPANotSpecified_setIsAgeRestrictedIsNeverInvoked() {
     val inMobiSdkWrapper = mock<InMobiSdkWrapper>()
     setCOPPAOnMobileAdsRequestConfiguration(
       RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED
@@ -86,7 +88,7 @@ class InMobiAdapterUtilsTest {
 
     InMobiAdapterUtils.setIsAgeRestricted(inMobiSdkWrapper)
 
-    verify(inMobiSdkWrapper).setIsAgeRestricted(false)
+    verify(inMobiSdkWrapper, never()).setIsAgeRestricted(any())
   }
 
   @Test