+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Version 11.0.0.0 (In progress)
 - Verified compatibility with InMobi Kotlin SDK version 11.0.0.
 - RewardedInterstitial support added.
+- Adapter does not modify InMobi isAgeRestricted if it is unspecified in the GMA SDK.
 
 #### Version 10.8.8.1
 - Removed class-level references to `Context` objects to help with memory leak issues.